affichage quad tree

+++ structs_gen.h
@@ -9,12 +9,12 @@ typedef struct {
 } Point2D;
 
 typedef struct {
-    int r;
-    int v;
-    int b;
+    float r;
+    float v;
+    float b;
 } Color3f;
 
-const Color3f COLOR3FNUL = {0,0,0};
+const Color3f COLOR3FNUL = {0.0,0.0,0.0};
 
 enum TypeRect {INDEFINI, DECORS, PERSONNAGE, PLATEFORME, PORTAIL};
